Crate ckb_chain_spec

The Chain Specification

By default, when simply running CKB, CKB will connect to the official public Nervos network.

In order to run a chain different to the official public one, with a config file specifying chain = "path" under [ckb].

Because the limitation of toml library, we must put nested config struct in the tail to make it serializable, details https://docs.rs/toml/0.5.0/toml/ser/index.html

Modules

consensus

default_params

Structs

ChainSpec

DepGroupResource

Genesis

GenesisCell

IssuedCell

Params

SatoshiGift

SystemCell

Enums

SpecError

SpecLoadError

Constants

OUTPUT_INDEX_DAO

OUTPUT_INDEX_SECP256K1_DATA

OUTPUT_INDEX_SECP256K1_BLAKE160_SIGHASH_ALL

OUTPUT_INDEX_SECP256K1_BLAKE160_MULTISIG_ALL

Functions

build_genesis_type_id_script

build_type_id_script

calculate_block_reward